Allow every day scheduler
