+++ app.rb
@@ -76,7 +76,13 @@ class GovukDependencies < Sinatra::Base
     return '[unauthorised]' unless params[:secret] == ENV['DEPENDAPANDA_SECRET']
 
     message_presenter = Presenters::Slack::FullMessage.new
-    UseCases::SendSlackMessages.new(message_presenter: message_presenter).execute(team: params.fetch(:team))
+    every_day_scheduler = UseCases::Slack::Schedulers::EveryDay.new
+
+    UseCases::SendSlackMessages.new(
+      message_presenter: message_presenter,
+      scheduler: every_day_scheduler
+    ).execute(team: params.fetch(:team))
+
     '[ok]'
   end
 end
