+++ lib/loader.rb
@@ -9,7 +9,8 @@ require_relative './use_cases/send_slack_messages'
 require_relative './use_cases/group_pull_requests_by_application'
 require_relative './use_cases/group_applications_by_team'
 require_relative './use_cases/group_pull_requests_by_gem'
-require_relative './use_cases/slack_message_scheduler'
+require_relative './use_cases/slack/schedulers/every_day'
+require_relative './use_cases/slack/schedulers/weekday'
 require_relative './gateways/pull_request'
 require_relative './gateways/pull_request_count'
 require_relative './gateways/slack_message'
