+++ lib/use_cases/send_slack_messages.rb
@@ -5,7 +5,7 @@ module UseCases
       team_usecase: UseCases::FetchTeams.new,
       pull_request_usecase: UseCases::FetchPullRequests.new,
       group_applications_by_team_usecase: UseCases::GroupApplicationsByTeam.new,
-      scheduler: UseCases::SlackMessageScheduler.new,
+      scheduler: UseCases::Slack::Schedulers::EveryDay.new,
       message_presenter:
     )
