+++ lib/use_cases/slack/schedulers/every_day.rb
@@ -1,0 +1,11 @@
+module UseCases
+  module Slack
+    module Schedulers
+      class EveryDay
+        def should_send_message?
+          true
+        end
+      end
+    end
+  end
+end
