--- lib/use_cases/slack_message_scheduler.rb
@@ -1,21 +1,0 @@
-module UseCases
-  class SlackMessageScheduler
-    def initialize(date_class: Date.new)
-      @date_class = date_class
-    end
-
-    def should_send_message?
-      return false if weekend?
-
-      true
-    end
-
-  private
-
-    attr_reader :date_class
-
-    def weekend?
-      date_class.saturday? || date_class.sunday?
-    end
-  end
-end
